+++ app/[locale]/(hotel)/hotel-list/page.jsx
@@ -87,8 +87,7 @@ const index = () => {
               <ListHotels />
             </div>
             <div id="restResults" className="tab-pane fade">
-              <h3>Menu 1</h3>
-              <p>Some content in menu 1.</p>
+              <ListHotels />
             </div>
           </div>
         </div>
